Run all the springs
This requires storing the dest. That's sad but it's more "correct" against future changes. Previously we only ran a subset of springs depending on whether we're in 1D or 2D mode. That subset happens to be the entire data, which was correct. But maybe in the future some items that aren't in some mode should still run